requirements.txt
aiohttp==3.11.18
cachetools==6.0.0
faiss_cpu==1.11.0
fastapi==0.115.12
langchain==0.3.25
langchain_community==0.3.24
numpy==2.2.6
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.3
rich==14.0.0
uvicorn==0.34.3
yandex_cloud_ml_sdk==0.10.0
